move spinner to common element
